infrastructure/requirements.txt
aws-cdk-lib>=2.100.0
constructs>=10.3.0
boto3>=1.34.0
botocore>=1.34.0